requirements.txt
django
django-widget-tweaks
django-active-link
django-macros
django-crispy-forms
django-select2
crispy-bootstrap4
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
fiobank
django-environ
django-tempus-dominus
django-tinymce
python-dateutil
django-mptt
django-polymorphic
djangorestframework
unidecode
django-crontab